shop/products link fix
